+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/rest/service/RestDeleteSpaceAction.java
@@ -136,7 +136,7 @@ public class RestDeleteSpaceAction extends BaseRestHandler {
             // Note: space is always DRAFT.
             // 1. Remove resources belonging to the space in Security Analytics.
             this.securityAnalyticsService.deleteSpaceResources(space);
-            // 2. Remove resources belonging to space in the .cti-* indices.
+            // 2. Remove resources belonging to space in the wazuh-threatintel-* indices.
             this.spaceService.deleteSpaceResources(space);
 
             // Re-generate the default policy for the space